feat: SearchAfterDependecy파일 분리
