+++ client/Targets/Presentation/Search/Implementations/Sources/Search/SearchResult/SearchAfterDashboard/SearchAfterDependency.swift
@@ -1,0 +1,23 @@
+//
+//  SearchAfterDependency.swift
+//  SearchImplementations
+//
+//  Created by 이준복 on 11/29/23.
+//  Copyright © 2023 codesquad. All rights reserved.
+//
+
+import ModernRIBs
+import DomainInterfaces
+
+protocol SearchAfterDashboardDependency: Dependency {
+    var searhResultSearchAfterUseCase: SearhResultSearchAfterUseCaseInterface { get }
+}
+
+protocol SearchAfterDashboardInteractorDependency: AnyObject {
+    var searhResultSearchAfterUseCase: SearhResultSearchAfterUseCaseInterface { get }
+}
+
+protocol SearchAfterDashboardRouterDependency {
+    var searchAfterStoryDashboardBuilder: SearchAfterStoryDashboardBuildable { get }
+    var searchAfterUserDashboardBuilder: SearchAfterUserDashboardBuildable { get }
+}
