+++ client/Targets/Presentation/Search/Implementations/Sources/Search/SearchResult/SearchAfterDashboard/SearchAfterStoryDashboard/SearchAfterStoryDashboardViewController.swift
@@ -135,7 +135,6 @@ extension SearchAfterStoryDashboardViewController: SearchAfterHeaderViewDelegate
     
 }
 
-
 extension SearchAfterStoryDashboardViewController: SearchAfterStoryViewDelegate {
     
     func searchAfterStoryViewDidTap(storyId: Int) {
